Refactor de_export.py, extract C++ function info to cxx_functions.py
Move everything related to information on C++ model functions to a separate module.

Related to #2306.

No changes in functionality.
